// src/maocars-client/schemas/comment.ts
/* eslint-disable */

/**
 * Generated by orval 🍺
 * Do not edit manually.
 * mao-cars-api
 */
import type { CommentType } from './commentType';
import type { UserShort } from './userShort';

export interface Comment {
  amount?: number;
  created_at: string;
  deleted_at?: string;
  id: string;
  text?: string;
  type: CommentType;
  updated_at?: string;
  upvotes: number;
  user: UserShort;
}